+++ src/components/StockList.tsx
@@ -1,13 +1,15 @@
-import { useAppSelector } from '../../store/hooks'
+import {useAtomValue, useSetAtom} from 'jotai';
+import { stocksAtom } from '../../store/atoms/';
+import { StockData } from '../../api';
 
 const StockList = () => {
-    const stockData = useAppSelector((state) => state.stocks.stockData);
-    console.log(stockData);
+    const stockData = useAtomValue(stocksAtom).stockData
+  
     return (
         <div>
-            <h1>Stock List</h1>
-            {stockData.map(stock=>{
-                return <p key={stock['01. symbol']}>{stock['01. symbol']}</p>
+            {stockData.length > 0 ? <h1>Stock List</h1> : <p>Select some tickers to analyze!</p>}
+            {stockData.map((stock:StockData)=>{
+                return <p key={stock.symbol}>{stock.symbol} : ${stock.alloc}</p>
             })}
         </div>
     )